Deletion now works as well, *mostly*.
Okay, so the thing is: add-text-property is END-exclusive.  Meaning
when you slap properties onto text regions the rightmost point won't
know about it.  Regions are END-inclusive.  If point is at the right
edge, it is still <=END (well, =END).  In other words: right-edge
should affect the text property lookup, but nothing regarding point
computation, which I fixed in this update.  However, there is YET
ANOTHER issue: if you delete a character and then move one character
ahead the active region will remain active, because the region
information is updated too late into the checking process.  What
remains is fixing the order of operations, which in itself will not
necessarily be trivial.