Greatly increased sanity of the backend.
Ignore the debug messages.

--point, --update-point
After lots of testing I came to the conclusion that it is unavoidable
to keep the darn point in check.  Trying to needlessly infer things
from region boundaries is unhelpful and just plain unintuitive.

--left-active-region-p
A conceptual wart from the get-go.  Now that font-lock is aware of the
point outside of font-lock: unneeded.

--recompute-region

An edge case prettify-symbols has problems dealing with is when
changes to the buffer cause the saved metadata to become stale.  This
is most likely a bug in ps and should probably reported.  Fixing the
issue is simple for single-line matches, but I have made the fix it's
own function to brace myself for future complications arising from
multi-line matches.